requirements.txt
discord.py
better-profanity